requirements.txt
Flask==3.0.0
Pillow==10.0.0
Pillow-Heif==0.15.0
opencv-python-headless==4.8.1
python-magic==0.4.27
python-magic-bin==0.4.14
numpy==1.24.3
werkzeug==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Limiter==3.5.1
Flask-CORS==4.0.0